replace img

+++ user/script.php
@@ -116,18 +116,24 @@ if(isset($_POST["btn_signup"])){
         }
     }
 
-#Replace Profile Image
-function replace_profile_img(){
+ #Replace Profile Image
+ if(isset($_POST["replace_submit"])){
+    $img = " ";
     $email = $_SESSION['email'];
-    if (($handle = fopen('../user/user_data.csv','r'))!== FALSE){
+    if (($handle = fopen('../user/account.csv','r'))!== FALSE){
         while (($data =  fgetcsv($handle,1000,",")) !== FALSE) {
             if ($data[1] == $email){
                     $img = $data[6];
+                    
             }
         }
     }
-    $img_name=$_FILES['profile_change']['name'];
-	$tmp_img_name=$_FILES['profile_change']['tmp_name'];
+    $img_name = $img;
+    if(file_exists($img)){
+        unlink($img);
+    }
+    $img_name=$_FILES['replace_image']['name'];
+	$tmp_img_name=$_FILES['replace_image']['tmp_name'];
     $img_name = $img;
     $folder='profile_img/';
 	move_uploaded_file($tmp_img_name,$folder.$img_name);
